dingest/ingest/requirements.txt
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.135.2
filelock==3.25.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
looseversion==1.3.0
lxml==6.0.2
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pathlib==1.0.1
pillow==12.1.1
pluggy==1.6.0
prov==2.1.1
puremagic==2.1.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.20.0
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
pyxnat==1.6.4
rdflib==7.6.0
reportlab==4.4.10
requests==2.33.0
scipy==1.17.1
simplejson==3.20.2
six==1.17.0
starlette==1.0.0
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
xlsxwriter==3.2.9
